Sequence Length Binning in Training
Implement that training supports creating batches out of bins instead of the whole training set. The bins are created so that token sequences of subjects, predicates and objects are as similar as possible, enabling shortening sequences before giving them to the model. A minimum number of triples in each bin can be given as an argument in the yaml.